Create deploy_lambda.sh
The deploy_lambda.sh script is a shell script used to deploy the Lambda function for the document management project on AWS. The script performs the following actions:

Create the deployment package: The first step in deploying a Lambda function is to create a deployment package that includes all the necessary code and dependencies. This script creates a deployment package by installing the required dependencies using pip, and then zipping the package directory.

Create the Lambda function: After creating the deployment package, the script uses the AWS CLI to create the Lambda function. The script specifies the function's name, runtime environment, execution role, handler function, timeout, memory size, and zip file.

Update the Lambda function: If the Lambda function already exists, the script updates the function's code with the new deployment package using the AWS CLI.

By using this script, you can quickly and easily deploy your Lambda function to AWS without having to manually perform each of these steps. This script can be executed from your local development environment or a build server as part of a continuous integration or continuous deployment (CI/CD) pipeline. It is recommended to store this script in version control and execute it as part of your deployment process.